# docs/swagger.yaml
basePath: /
definitions:
  dto.Customer:
    properties:
      cpf:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - cpf
    - email
    - name
    type: object
...
  dto.CustomerResponse:
    properties:
      id:
        type: integer
    type: object
...
host: localshot:3210
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API for the Tech1 Fiap Project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tech1 API Docs
  version: "1.0"
paths:
  /api/admin/customers/{id}:
    put:
      consumes:
      - application/json
      description: Update customer
      parameters:
      - description: "12"
        in: path
        name: id
        required: true
        type: integer
      - description: customer
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.Customer'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Customer has required fields
        "404":
          description: Customer not found
      summary: Update customer
      tags:
      - Customer
  /api/customers/{id}:
    get:
      consumes:
      - application/json
      description: Get customer by ID
      parameters:
      - description: "12"
        in: path
        name: Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Customer'
        "404":
          description: Customer not found
      summary: Get customer by ID
      tags:
      - Customer
...
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create new customer. This process is not required to make an order
      parameters:
      - description: customer
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CustomerResponse'
        "400":
          description: Customer has required fields
        "409":
          description: This Customer is already added
      summary: Create new customer
      tags:
      - Customer
swagger: "2.0"